requirements.txt
llama-cpp-python>=0.2.23
pypdf>=3.17.0
gradio>=3.44.0
python-dotenv>=1.0.0